[fix] free-timeを取得する関数を修正
