+++ src/internal/handlers/freetimes/freetimes.go
@@ -2,15 +2,15 @@ package gateway
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"net/http"
 	"src/internal/config"
 	"src/internal/entity"
 	"src/internal/entity/validation"
 	"src/internal/repository"
-	"src/internal/utils/num"
+	"src/internal/test/time"
 	"src/internal/utils/strings"
-	"src/internal/utils/time"
 	"strconv"
 
 	"github.com/jmoiron/sqlx"
@@ -105,7 +105,7 @@ func CreateFreeTime(ctx context.Context, db *sqlx.DB) echo.HandlerFunc {
 		// 入力された日付のdate-free-timeを取得する
 		dateFreeTime, err := repository.GetDateFreeTime(ctx, db, userID, year, month, day)
 		// 指定した日付のdate-free-timeが存在しなかった場合
-		if err == entity.ErrNoDateFreeTimeFound {
+		if errors.Is(err, entity.ErrNoDateFreeTimeFound) {
 			// DateFreeTime構造体で値を設定する
 			dateFreeTime = &entity.DateFreeTime{
 				UserID: userID,
@@ -126,12 +126,15 @@ func CreateFreeTime(ctx context.Context, db *sqlx.DB) echo.HandlerFunc {
 			})
 		}
 
-		// free-timeが作成できるかどうか
-		checkResult = time.IsCreateFreeTime(startFreeTimeHour, startFreeTimeMinute, endFreeTimeHour, endFreeTimeMinute, dateFreeTime)
-		if !checkResult {
-			return c.Render(http.StatusOK, "create-free-time", map[string]interface{}{
-				"error_message": entity.ERR_ALREADY_FREE_TIME_EXISTS,
-			})
+		// 指定した日付のfree-timeが存在した場合
+		if dateFreeTime.FreeTimes != nil {
+			// free-timeが作成できるかどうか
+			checkResult = time.IsCreateFreeTime(startFreeTimeHour, startFreeTimeMinute, endFreeTimeHour, endFreeTimeMinute, dateFreeTime)
+			if !checkResult {
+				return c.Render(http.StatusOK, "create-free-time", map[string]interface{}{
+					"error_message": entity.ERR_ALREADY_FREE_TIME_EXISTS,
+				})
+			}
 		}
 
 		// free-timeを作成するための値を構造体に格納する
@@ -151,10 +154,6 @@ func CreateFreeTime(ctx context.Context, db *sqlx.DB) echo.HandlerFunc {
 		}
 
 		jpWeekday := time.GetWeekdayByDate(year, month, day)
-		startFreeTimeHourStr = num.NumToFormattedString(startFreeTimeHour)
-		startFreeTimeMinuteStr = num.NumToFormattedString(startFreeTimeMinute)
-		endFreeTimeHourStr = num.NumToFormattedString(endFreeTimeHour)
-		endFreeTimeMinuteStr = num.NumToFormattedString(endFreeTimeMinute)
 
 		successMessage := fmt.Sprintf("%s/%s/%s（%s）%s:%s〜%s:%sでfree-timeを作成しました。", yearStr, monthStr, dayStr, jpWeekday, startFreeTimeHourStr, startFreeTimeMinuteStr, endFreeTimeHourStr, endFreeTimeMinuteStr)
