+++ src/internal/web/static/css/base.css
@@ -15,6 +15,8 @@ input,textarea,select,button {
   font-family: 'Andale Mono','Sitka Heading','Avenir',sans-serif;
   color: var(--BLACK);
   border: none;
+  outline: none
+
 }
 
 a {
@@ -27,6 +29,7 @@ a,a:hover,a:visited{
 ul {
   list-style-type: none;
 }
+
 
 /* =================== FlexBox ====================== */
