+++ src/internal/web/template/create-free-time.html
@@ -109,10 +109,10 @@
               <div class="select-time-start-hour-area flex">
                 <select class="select-time-start-hour font-17" name="start-free-time-hour">
                   <option value="">開始時刻</option>
-                  <option value="6">06</option>
-                  <option value="7">07</option>
-                  <option value="8">08</option>
-                  <option value="9">09</option>
+                  <option value="06">06</option>
+                  <option value="07">07</option>
+                  <option value="08">08</option>
+                  <option value="09">09</option>
                   <option value="10">10</option>
                   <option value="11">11</option>
                   <option value="12">12</option>
@@ -140,10 +140,10 @@
               <div class="select-time-end-hour-area">
                 <select class="select-time-end-hour font-17" name="end-free-time-hour">
                   <option value="">終了時刻</option>
-                  <option value="6">06</option>
-                  <option value="7">07</option>
-                  <option value="8">08</option>
-                  <option value="9">09</option>
+                  <option value="06">06</option>
+                  <option value="07">07</option>
+                  <option value="08">08</option>
+                  <option value="09">09</option>
                   <option value="10">10</option>
                   <option value="11">11</option>
                   <option value="12">12</option>
